requirements.txt
celery==5.5.2
Django==5.2.1
flake8==7.2.0
isort==6.0.1
redis==6.1.0
requests==2.28.2
djangorestframework~=3.16.0
drf-yasg~=1.21.10